+++ lib/src/properties/bottom_sheet_props.dart
@@ -1,4 +1,6 @@
-import 'package:flutter/material.dart';
+import 'dart:ui';
+
+import 'package:flutter/cupertino.dart';
 
 class BottomSheetProps {
   final ShapeBorder? shape;
@@ -25,26 +27,24 @@ class BottomSheetProps {
 }
 
 class CupertinoBottomSheetProps {
-  final ShapeBorder? shape;
-  final BoxConstraints? constraints;
-  final Color? backgroundColor;
-  final Clip clipBehavior;
-  final AnimationController? transitionAnimationController;
-  final bool enableDrag;
-  final double? elevation;
-  final bool? showDragHandle;
-  final AnimationStyle? sheetAnimationStyle;
+  final bool useRootNavigator;
+  final ImageFilter? filter;
+  final bool barrierDismissible;
+  final bool semanticsDismissible;
+  final RouteSettings? routeSettings;
+  final bool isSurfacePainted;
+  final Offset? anchorPoint;
+  final Color barrierLabel;
 
   const CupertinoBottomSheetProps({
-    this.elevation,
-    this.shape,
-    this.backgroundColor,
-    this.transitionAnimationController,
-    this.enableDrag = true,
-    this.clipBehavior = Clip.none,
-    this.constraints,
-    this.showDragHandle,
-    this.sheetAnimationStyle,
+    this.anchorPoint,
+    this.routeSettings,
+    this.isSurfacePainted = true,
+    this.filter,
+    this.barrierDismissible = true,
+    this.semanticsDismissible = false,
+    this.useRootNavigator = false,
+    this.barrierLabel = kCupertinoModalBarrierColor,
   });
 }